+++ Rdmp.Core/CohortCreation/ICompileable.cs
@@ -29,6 +29,8 @@ namespace Rdmp.Core.CohortCreation
         event EventHandler StateChanged;
         Exception CrashMessage { get; set; }
 
+        string Log { get; set; }
+
         int FinalRowCount { get; set; }
         int? CumulativeRowCount { get; set; }